backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
datasets>=2.14.0
Pillow>=10.0.0
numpy>=1.24.0
requests>=2.31.0
diffusers>=0.21.0
realesrgan>=0.3.0
basicsr>=1.4.0
